+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Keep it human-readable, your future self will thank you!
 ### Added
 
 - Adding the user recipe in the dataset PR #59.
+- Add `multi_dates_match` action in create.
 
 ### Changed